Fix Livewire Browser Plugin component updates (#136)
With the Livewire Browser Plugin you can modify component properties, which generate an AJAX call back to the server (similar to `form.set('address.country_id', 'NL')` via the MageWire JS API). However, in that case, the request parameters like the `fingerprint` are sent as part of a JSON body, and not with request parameters. This PR simply supports both.